refatoracao

+++ src/modules/user/user.service.ts
@@ -2,13 +2,16 @@ import { Injectable } from '@nestjs/common'
 import { InjectRepository} from '@nestjs/typeorm'
 
 import { CryptUtilityService } from '@shared/bcrypt/bcrypt.service'
-import { OK, NotFoundExceptions } from '@service/exception'
+import { OK, NotFoundExceptions, ConflictExceptions } from '@service/exception'
 import { code, message } from '@shared/enum'
+
+import { FirebaseValidate } from '@modules/firebase/firebase.validate'
 
 import { UserValidate } from './user.validate'
 import { UserRepository } from './user.repository'
 import { PerfilUserMapper, CheckUserExistsByEmailMapper } from './mapper'
 import { CreateNewUserDto, UpdateUserDto } from './dto'
+
 @Injectable()
 export class UserService {
 
@@ -16,6 +19,7 @@ export class UserService {
     @InjectRepository(UserRepository) 
     private readonly repository: UserRepository, 
     private validate:UserValidate,
+    private validateFirebase:FirebaseValidate,
     private crypt:CryptUtilityService,
     private perfilUserMapper:PerfilUserMapper,
     private checkUserExistsByEmailMapper:CheckUserExistsByEmailMapper
@@ -35,8 +39,12 @@ export class UserService {
     )
   }
 
-  public async findAll() {
+  public async findAll(token:string) {
+    let body = await this.validateFirebase.isToken(token)
+    let decoded = await this.validateFirebase.validateTokenByFirebase(body)
+
     const res = await this.repository.find();
+    
     if(Object.keys(res).length == 0){
       throw new NotFoundExceptions({
         code:code.NOT_FOUND_USER,
@@ -47,13 +55,35 @@ export class UserService {
     return new OK(dto)
   }
 
-  public async getUserByUid(uid:string) {
+  public async getUserByUid(uid:string, token:string) {
+    let body = await this.validateFirebase.isToken(token)
+    let decoded = await this.validateFirebase.validateTokenByFirebase(body)
+
+    if(decoded.uid != uid){
+      return await new ConflictExceptions({
+        code:code.UID_INVALID,
+        message:message.UID_INVALID,
+        description:message.UID_INVALID_CONFLICT_TOKEN_DESCRIPTION
+      })
+    }
+
     const res = await this.validate.getUserByUid(uid)
     const dto = this.perfilUserMapper.toDto(res)
     return new OK([dto])
   }
 
-  public async getUserByEmail(email:string) {
+  public async getUserByEmail(email:string, token:string) {
+    let body = await this.validateFirebase.isToken(token)
+    let decoded = await this.validateFirebase.validateTokenByFirebase(body)
+
+    if(decoded.email != email){
+      return await new ConflictExceptions({
+        code:code.EMAIL_INVALID,
+        message:message.EMAIL_INVALID,
+        description:message.EMAIL_INVALID_CONFLICT_TOKEN_DESCRIPTION
+      })
+    }
+
     const res = await this.validate.getUserByEmail(email)
     const dto = this.perfilUserMapper.toDto(res)
     return new OK([dto])
@@ -65,7 +95,18 @@ export class UserService {
     return new OK([dto])
   }
 
-  public async updateUserByUid(uid:string, user:UpdateUserDto) {
+  public async updateUserByUid(uid:string, user:UpdateUserDto, token:string) {
+    let body = await this.validateFirebase.isToken(token)
+    let decoded = await this.validateFirebase.validateTokenByFirebase(body)
+
+    if(decoded.uid != uid){
+      return await new ConflictExceptions({
+        code:code.UID_INVALID,
+        message:message.UID_INVALID,
+        description:message.UID_INVALID_CONFLICT_TOKEN_DESCRIPTION
+      })
+    }
+
     const { id } = await this.validate.getUserByUid(uid)
     await this.validate.updateUserByUid(id, user)
     const res = await this.validate.getUserByUid(uid)
@@ -73,13 +114,27 @@ export class UserService {
     return new OK([dto], code.USER_UPDATED, message.USER_UPDATED) 
   }
 
-  public async deleteUserByUid(uid:any) {
+  public async deleteUserByUid(uid:any, token:string) {
+    let body = await this.validateFirebase.isToken(token)
+    let decoded = await this.validateFirebase.validateTokenByFirebase(body)
+
+    if(decoded.uid != uid){
+      return await new ConflictExceptions({
+        code:code.UID_INVALID,
+        message:message.UID_INVALID,
+        description:message.UID_INVALID_CONFLICT_TOKEN_DESCRIPTION
+      })
+    }
+
     const { id } = await this.validate.getUserByUid(uid)
     await this.validate.deleteUserByUid(id)
     return new OK([], code.DELETED_SUCCESSFULLY, message.DELETED_SUCCESSFULLY) 
   }
 
-  public async deleteTodosUsuarios() {
+  public async deleteTodosUsuarios(token:string) {
+    let body = await this.validateFirebase.isToken(token)
+    let decoded = await this.validateFirebase.validateTokenByFirebase(body)
+
     await this.repository.deleteTodosUsuarios();
     return new OK([],code.DELETED_SUCCESSFULLY, message.DELETED_SUCCESSFULLY) 
   }
